Fix WalletUnlocker client
