+++ src/types/rpc/index.ts
@@ -5,5 +5,6 @@ export * from './ln-rpc';
 export * from './router-rpc';
 export * from './sign-rpc';
 export * from './wallet-rpc';
+export * from './wallet-unlocker-rpc';
 export * from './watchtower-rpc';
 export * from './wt-client-rpc';
